LayoutTests/inspector/console/console-messages-stack-traces-expected.txt

CONSOLE MESSAGE: line 17: log
CONSOLE MESSAGE: line 18: info
CONSOLE MESSAGE: line 19: warn
CONSOLE MESSAGE: line 20: debug
CONSOLE MESSAGE: line 21: error
CONSOLE MESSAGE: line 22: assert
CONSOLE MESSAGE: line 23: trace
CONSOLE MESSAGE: line 24: 1,2,3
CONSOLE MESSAGE: line 25: [object HTMLBodyElement]
CONSOLE MESSAGE: line 31: MARKER: EXPECT NO STACK TRACES
CONSOLE MESSAGE: line 35: MARKER: DONE
Tests that console-api messages contain stack traces.

SUCCESS
